requirements.txt
absl-py==0.9.0
astor==0.8.1
cycler==0.10.0
gast==0.3.3
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.2.1
matplotlib==3.2.1
numpy==1.18.1
opencv-python==4.2.0.32
Pillow==7.0.0
pkg-resources==0.0.0
protobuf==3.11.3
pyparsing==2.4.6
python-dateutil==2.8.1
six==1.14.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
Werkzeug==1.0.0
wrapt==1.12.1
